refactor: Rename routes from /process-events to /records
